* commit 'e25772436509322136b960af8716fc443f67ef2e':
  Se agrega columna de total
  Se quita la conversion a entero de la cantidad completa
  Se cambia ordenamiento de tarjetas

+++ services/sapadapter-service/Omicron.SapAdapter.Repository/Model/JoinsModels/CompleteFormulaWithDetalle.cs
@@ -121,7 +121,7 @@ namespace Omicron.SapAdapter.Entities.Model.JoinsModels
         /// Gets or sets Code.
         /// </summary>
         /// <value>The code.</value>
-        public int CompleteQuantity { get; set; }
+        public decimal CompleteQuantity { get; set; }
 
         /// <summary>
         /// Gets or sets Code.
